build: new set_utils.cmake for various handy "set()" wrappers
New set_utils.cmake has some handy new utilities: set_cache() is much
like the built-in set() when making a cache variable, and set_option()
is much like build-in option(). The biggest difference is that both
allow an environment variable of the same name, if it exists, to
supply the default value. This is something that cmake does with many
of its own controls, like CMAKE_BUILD_TYPE, but does not make any
provision for built-in set() or option() let users do it.

Signed-off-by: Larry Gritz <lg@larrygritz.com>