ci: For Windows CI, build only release of vcpkg packages
This cuts down CI time, as before we were building both release and
debug versions of the packages.

The dependency building step of our CI typically took around 6.5-7
minutes previously, and with this patch, now typically is about 5
minutes and sometimes slightly less.

Thanks so much to @cedrik-fuoco who pointed me at this feature of
vcpkg which had eluded me for a long time!

Signed-off-by: Larry Gritz <lg@larrygritz.com>