admin: Document my git-cliff workflow for release notes
For the last year or so, I have been using git-cliff to generate the
first draft of release notes. It gets me 90% of the way there for the
monthly patch releases, perhaps more like 30% (but a very good
starting point) for the big annual major releases.

In this PR, I document my process and check in the toml file that
provides the git cliff configuration that gets mostly what we need.

Signed-off-by: Larry Gritz <lg@larrygritz.com>