cleanup: Make span::size() return size_t
With 3.0 coming soon, it's time to break compatibility by switching
OIIO::span::size() to return an unsized value, matching C++20
std::span.

The history is that when we first added OIIO::span, the C++ draft
proposal for std::span included a signed size() result, but by the
time it got finalized into C++20, it was changed to unsigned to match
all the other std containers, but it was too late for us to switch
without breaking API compatibility rather seriously.

Signed-off-by: Larry Gritz <lg@larrygritz.com>