api: imagebuf.h add deprecation warnings to deprecated things
* Add deprecation warnings to the varieties of ImageBuf constructor and
  `reset()` that don't take subimage and miplevel parameters. The equivalent
  is to just pass `0` for both of those parameters. The warned constructors
  have been marked as deprecated since OIIO 2.2.
* The misspelled `ImageBuf::make_writeable()` has been given deprecation
  warnings. Since OIIO 2.2, the correct spelling is `make_writable`.
* The `ImageBuf::error()` method that uses printf-style formatting now has
  deprecation warnings. Use `ImageBuf::errorfmt()` instead.
* The `ImageBuf::interppixel_NDC_full()` method, which has been marked as
  deprecated since OIIO 1.5, now has deprecation warnings. Use
  `interppixel_NDC()` instead.

Undeprecate the name+spec varieties of constructors. They were
documented as deprecated but warnings were never added. In doing this
set of changes, I realized that they were used for something useful
after all -- it makes it easy to remember the name of a new buffer
you're making not from a file, so you can later `write()` it without
needing to know the filename at that time.  I also added a new
`set_name()` method to do it directly for times when you don't want to
rely on the constructor or want to change the name.

Signed-off-by: Larry Gritz <lg@larrygritz.com>