Finish removing Imath include (#4371)
https://github.com/lgritz/OpenImageIO/commit/4cb956b0a0a123376e468a5b5e6e188178ec077e
lost this removal when resolving a merge conflict. This properly removes
it.

https://github.com/AcademySoftwareFoundation/OpenImageIO/pull/4076
implemented matrix44::inverse() so that we could stop including Imath.h.
It removed the Imath.h include in the master/dev-2.6 branch but the
dev-2.5 branch kept the include, likely because of an improperly
resolved merge conflict. This PR removes the include like in master.

No tests are needed. As long as it compiles, this should be good. I
verified this builds on arm64.

Signed-off-by: Thiago Ize <10453105+ThiagoIze@users.noreply.github.com>
Co-authored-by: Thiago Ize <10453105+ThiagoIze@users.noreply.github.com>