api: use shared_ptr for ImageCache and TextureSystem
Change IC::create() and TS::create() to return shared_ptr instead of
a raw pointer. This cleans up a lot of lingering lifetime management
issue of these classes. The switch to 3.0 is an opportunity to make
breaking changes to these APIs.

For the sake of downstream users, we define
OIIO_IMAGECACHE_CREATE_SHARED and OIIO_TEXTURESYSTEM_CREATE_SHARED to
signal that the new APIs are present. Some very minor changes may be
needed at the call site.

Signed-off-by: Larry Gritz <lg@larrygritz.com>