build(deps): Raise libheif minimum to 1.11
This version dates from early 2021 (so still supporting back 3.5 years).

The reason for the bump is to simplify a bit on our part and avoid
older versions known to be broken or lacking certain features.

This is for master/2.6/3.0 only, will not be backported to 2.5, since
we never raise minimum dependencies in already-released branches.

Signed-off-by: Larry Gritz <lg@larrygritz.com>