fix(iba): IBA::transpose() didn't set output image's format to input
Because IBA::transpose changes resolution, it doesn't pass the input
image to IBAPrep(), and there was a special case there that always set
destination to float if there were no source images.

So change transpose to specifically request a format, to be the same
as the input image.

And that required a tweak to IBAPrep allowing us to specify such a
thing (there was already an option that meant "force it to float") but
not the general case of an arbitrary type.

Signed-off-by: Larry Gritz <lg@larrygritz.com>