feat(jpeg): Support encoding/decoding arbitrary metadata as comments
This is needed to port Blender's current JPEG IO code to using OIIO,
but is also a useful feature to have in general.

For reading, the code tries to parse comments as colon-separated key-value
pairs and sets metadata accordingly. For writing, this needs to be explicitly
enabled by setting jpeg:com_attributes to 1 in order to avoid accidentally
bloating files for existing applications.

Signed-off-by: Lukas Stockner <lukas@lukasstockner.de>