fix(openexr): modernize dwa compression level setting
Starting with OpenEXR 3.1.3, the preferred API call for setting
DWAA/DWAB compression level has changed. We never changed the OIIO
side. Luckily, OpenEXR seems to have kept respecting the old API calls
we were making (passing as an attribute). But this modernizes the
approach, now that we don't have OpenEXR 2.x suppor to maintain.

But in the process, we also fixed a bug! Turns out we weren't
propagating the compression properly -- if the compression name
was plain "dwaa", we ignored any "openexr:dwaCompressionLevel"
attribute, effectively.

Signed-off-by: Larry Gritz <lg@larrygritz.com>